feat: Warning messages on wrong practices (#314)
* feat: Warning messages on wrong practices

* test: Group

* feat: Add warnings
