+++ src/lib.rs
@@ -28,16 +28,13 @@ mod from_str;
 mod generator;
 #[cfg(feature = "json_schema")]
 mod json_schema;
+mod linter;
 pub mod lock;
 #[cfg(feature = "json_schema")]
 use schemars::gen::*;
 pub use {
-    context::*,
-    deserialize::*,
-    dofigen_struct::*,
-    errors::*,
-    extend::*,
-    generator::{GenerationContext, MessageLevel},
+    context::*, deserialize::*, dofigen_struct::*, errors::*, extend::*,
+    generator::GenerationContext, linter::*,
 };
 
 #[cfg(all(feature = "strict", feature = "permissive"))]
@@ -77,8 +74,12 @@ const FILE_HEADER_COMMENTS: [&str; 2] = [
 ///     "# syntax=docker/dockerfile:1.7\n# This file is generated by Dofigen v0.0.0\n# See https://github.com/lenra-io/dofigen\n\n# runtime\nFROM ubuntu AS runtime\nUSER 1000:1000\n"
 /// );
 /// ```
+#[deprecated(
+    since = "2.2.0",
+    note = "Please use `GenerationContext::generate_dockerfile` from `dofigen_lib` instead"
+)]
 pub fn generate_dockerfile(dofigen: &Dofigen) -> Result<String> {
-    GenerationContext::from(dofigen).generate_dockerfile(dofigen)
+    GenerationContext::from(dofigen.clone()).generate_dockerfile()
 }
 
 /// Generates the .dockerignore file content from an Dofigen struct.
@@ -136,31 +137,14 @@ pub fn generate_dockerfile(dofigen: &Dofigen) -> Result<String> {
 ///     "# This file is generated by Dofigen v0.0.0\n# See https://github.com/lenra-io/dofigen\n\n**\n!/src\n/src/*.test.rs\n"
 /// );
 /// ```
+#[deprecated(
+    since = "2.2.0",
+    note = "Please use `GenerationContext::generate_dockerignore` from `dofigen_lib` instead"
+)]
 pub fn generate_dockerignore(dofigen: &Dofigen) -> String {
-    let mut content = String::new();
-
-    for line in FILE_HEADER_COMMENTS {
-        content.push_str("# ");
-        content.push_str(line);
-        content.push_str("\n");
-    }
-    content.push_str("\n");
-
-    if !dofigen.context.is_empty() {
-        content.push_str("**\n");
-        dofigen.context.iter().for_each(|path| {
-            content.push_str("!");
-            content.push_str(path);
-            content.push_str("\n");
-        });
-    }
-    if !dofigen.ignore.is_empty() {
-        dofigen.ignore.iter().for_each(|path| {
-            content.push_str(path);
-            content.push_str("\n");
-        });
-    }
-    content
+    GenerationContext::from(dofigen.clone())
+        .generate_dockerignore()
+        .unwrap()
 }
 
 /// Generates the effective Dofigen content from a Dofigen struct.
