+++ tests/lib_test.rs
@@ -9,7 +9,9 @@ fn yaml_to_dockerfile_empty() {
         .parse_from_string(yaml)
         .map_err(Error::from)
         .unwrap();
-    let dockerfile: String = generate_dockerfile(&dofigen).unwrap();
+
+    let mut generation_context = GenerationContext::from(dofigen);
+    let dockerfile: String = generation_context.generate_dockerfile().unwrap();
 
     assert_eq_sorted!(
         dockerfile,
@@ -23,7 +25,7 @@ USER 1000:1000
 "#
     );
 
-    let dockerignore: String = generate_dockerignore(&dofigen);
+    let dockerignore: String = generation_context.generate_dockerignore().unwrap();
 
     assert_eq_sorted!(
         dockerignore,
@@ -33,12 +35,12 @@ USER 1000:1000
 
 #[test]
 #[cfg(feature = "permissive")]
-fn yaml_to_dockerfile_complexe() {
+fn yaml_to_dockerfile_complex() {
     let yaml = r#"
 builders:
   builder:
     fromImage: ekidd/rust-musl-builder
-    user: rust
+    user: 1000
     add: "."
     run:
     - ls -al
@@ -72,7 +74,8 @@ ignores:
         .parse_from_string(yaml)
         .map_err(Error::from)
         .unwrap();
-    let dockerfile: String = generate_dockerfile(&dofigen).unwrap();
+    let mut generation_context = GenerationContext::from(dofigen);
+    let dockerfile: String = generation_context.generate_dockerfile().unwrap();
 
     assert_eq_sorted!(
         dockerfile,
@@ -83,12 +86,12 @@ ignores:
 # builder
 FROM ekidd/rust-musl-builder AS builder
 COPY \
-    --chown=rust \
+    --chown=1000 \
     --link \
     "." "./"
-USER rust
+USER 1000
 RUN \
-    --mount=type=cache,target=/usr/local/cargo/registry,sharing=locked \
+    --mount=type=cache,target=/usr/local/cargo/registry,uid=1000,sharing=locked \
     <<EOF
 ls -al
 cargo build --release
@@ -120,9 +123,11 @@ CMD ["/fwatchdog"]
 "#
     );
 
-    let dockerignore: String = generate_dockerignore(&dofigen);
+    let dockerignore: String = generation_context.generate_dockerignore().unwrap();
 
     assert_eq_sorted!(dockerignore, "# This file is generated by Dofigen v0.0.0\n# See https://github.com/lenra-io/dofigen\n\ntarget\ntest\n");
+
+    assert_eq_sorted!(generation_context.get_lint_messages(), vec![]);
 }
 
 #[ignore = "Flatten enum variant alias problem: https://github.com/serde-rs/serde/issues/2188"]
@@ -217,7 +222,8 @@ run:
 "#;
 
     let dofigen: Dofigen = DofigenContext::new().parse_from_string(yaml).unwrap();
-    let dockerfile: String = generate_dockerfile(&dofigen).unwrap();
+    let mut generation_context = GenerationContext::from(dofigen);
+    let dockerfile: String = generation_context.generate_dockerfile().unwrap();
 
     assert_eq_sorted!(
         dockerfile,
@@ -398,7 +404,7 @@ cache:
 builders:
   builder:
     fromImage: ekidd/rust-musl-builder
-    user: rust
+    user: 1000
     add: "."
     run: cargo build --release
     cache: /usr/local/cargo/registry
@@ -412,7 +418,8 @@ cache: /tmp
 "#;
 
     let dofigen: Dofigen = DofigenContext::new().parse_from_string(yaml).unwrap();
-    let dockerfile: String = generate_dockerfile(&dofigen).unwrap();
+    let mut generation_context = GenerationContext::from(dofigen);
+    let dockerfile: String = generation_context.generate_dockerfile().unwrap();
 
     assert_eq_sorted!(
         dockerfile,
@@ -423,12 +430,12 @@ cache: /tmp
 # builder
 FROM ekidd/rust-musl-builder AS builder
 COPY \
-    --chown=rust \
+    --chown=1000 \
     --link \
     "." "./"
-USER rust
+USER 1000
 RUN \
-    --mount=type=cache,target=/usr/local/cargo/registry,sharing=locked \
+    --mount=type=cache,target=/usr/local/cargo/registry,uid=1000,sharing=locked \
     cargo build --release
 
 # runtime
@@ -444,6 +451,8 @@ RUN \
     echo "coucou"
 "#
     );
+
+    assert_eq_sorted!(generation_context.get_lint_messages(), vec![]);
 }
 
 #[test]
@@ -521,7 +530,8 @@ context:
 "#;
 
     let dofigen: Dofigen = DofigenContext::new().parse_from_string(yaml).unwrap();
-    let dockerfile: String = generate_dockerfile(&dofigen).unwrap();
+    let mut generation_context = GenerationContext::from(dofigen);
+    let dockerfile: String = generation_context.generate_dockerfile().unwrap();
 
     assert_eq_sorted!(
         dockerfile,
@@ -556,6 +566,8 @@ ENTRYPOINT ["/bin/dofigen"]
 CMD ["--help"]
 "#
     );
+
+    assert_eq_sorted!(generation_context.get_lint_messages(), vec![]);
 }
 
 #[ignore = "Not managed yet by serde because of multilevel flatten: https://serde.rs/field-attrs.html#flatten"]
